+++ server/src/main/webapp/WEB-INF/rails/app/assets/javascripts/application.js
@@ -78,6 +78,6 @@
 //= require "micro_content_popup.js"
 //= require "console_log_tailing.js"
 //= require "js-routes"
-//= require "xhr_promise.js"
+//= require "promise-finally-override.js"
 //= require_directory .
